+++ mongodb/src/lib.rs
@@ -39,6 +39,7 @@ use replicante_util_tracing::tracer;
 mod agent;
 mod config;
 mod errors;
+mod interface;
 mod metrics;
 mod rs_status;