+++ mongodb/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - **BREAKING**: Rename MongoDB Agent binary
 - **BREAKING**: Rework configuration using [serde](https://docs.rs/serde)
 - **BREAKING**: Use [error_chain](https://docs.rs/error-chain) for errors
+- Move MongoDB interactions into a trait
 
 ## 0.1.0 - 2018-06-28
 ### Added